functions/requirements.txt
azure-functions
coolname
